requirements.txt
dnspython==2.7.0
kafka-python-ng==2.2.3
py4j==0.10.9.7
pymongo==4.10.1
pyspark==3.5.4
